+++ go/coinstacks/thorchain/pulumi/index.ts
@@ -1,7 +1,7 @@
 import { parse } from 'dotenv'
 import { readFileSync } from 'fs'
 import * as k8s from '@pulumi/kubernetes'
-import { daemonService, deployStatefulService, getConfig, Service } from '../../../pulumi/src'
+import { createService, deployStatefulService, getConfig, Service } from '../../../pulumi/src'
 
 type Outputs = Record<string, any>
 
@@ -39,12 +39,29 @@ export = async (): Promise<Outputs> => {
     const services: Record<string, Service> = {}
 
     if (config.statefulService.daemon) {
-      services.daemon = daemonService({
+      services.daemon = createService({
         asset,
         config: config.statefulService.daemon,
         dataDir: '/root', // default home path for thornode docker image
         env: {'CHAIN_ID': `thorchain-${config.network}-v1`, 'NET': config.network},
+        name: 'daemon',
         ports: {'daemon-api': 1317, 'daemon-rpc': 27147}
+      })
+    }
+
+    if (config.statefulService.timescaledb) {
+      services.timescaledb = createService({
+        asset,
+        config: config.statefulService.timescaledb,
+        dataDir: '/var/lib/postgresql/data',
+        env: {
+          'POSTGRES_DB': 'midgard',
+          'POSTGRES_USER': 'midgard',
+          'POSTGRES_PASSWORD': 'password',
+          'PGDATA': '/var/lib/postgresql/data/pgdata'
+        },
+        name: 'timescaledb',
+        ports: {'postgres': 5432}
       })
     }
