+++ go/coinstacks/thorchain/timescaledb/readiness.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+pg_isready -U midgard
+
+exit echo $?
